+++ praw/util/token_manager.py
@@ -3,13 +3,13 @@
 There should be a 1-to-1 mapping between an instance of a subclass of
 :class:`.BaseTokenManager` and a :class:`.Reddit` instance.
 
-A few trivial token manager classes are provided here, but it is expected that PRAW
-users will create their own token manager classes suitable for their needs.
+A few proof of concept token manager classes are provided here, but it is expected that
+PRAW users will create their own token manager classes suitable for their needs.
 
-See ref:`using_refresh_tokens` for examples on how to leverage these classes.
+See :ref:`using_refresh_tokens` for examples on how to leverage these classes.
 
 """
-
+import sqlite3
 from abc import ABC, abstractmethod
 
 
@@ -59,7 +59,19 @@ class BaseTokenManager(ABC):
 
 
 class FileTokenManager(BaseTokenManager):
-    """Provides a trivial single-file based token manager."""
+    """Provides a single-file based token manager.
+
+    It is expected that the file with the initial ``refresh_token`` is created prior to
+    use.
+
+    .. warning::
+
+        The same ``file`` should not be used by more than one instance of this class
+        concurrently. Doing so may result in data corruption. Consider using
+        :class:`.SQLiteTokenManager` if you want more than one instance of PRAW to
+        concurrently manage a specific ``refresh_token`` chain.
+
+    """
 
     def __init__(self, filename):
         """Load and save refresh tokens from a file.
@@ -80,3 +92,98 @@ class FileTokenManager(BaseTokenManager):
         if authorizer.refresh_token is None:
             with open(self._filename) as fp:
                 authorizer.refresh_token = fp.read().strip()
+
+
+class SQLiteTokenManager(BaseTokenManager):
+    """Provides a SQLite3 based token manager.
+
+    Unlike, :class:`.FileTokenManager`, the initial database need not be created ahead
+    of time, as it'll automatically be created on first use. However, initial
+    ``refresh_tokens`` will need to be registered via :meth:`.register` prior to use.
+    See :ref:`sqlite_token_manager` for an example of use.
+
+    .. warning::
+
+        This class is untested on Windows because we encountered file locking issues in
+        the test environment.
+
+    """
+
+    def __init__(self, database, key):
+        """Load and save refresh tokens from a SQLite database.
+
+        :param database: The path to the SQLite database.
+        :param key: The key used to locate the ``refresh_token``. This ``key`` can be
+            anything. You might use the ``client_id`` if you expect to have unique
+            ``refresh_tokens`` for each ``client_id``, or you might use a Redditor's
+            ``username`` if you're manage multiple users' authentications.
+
+        """
+        super().__init__()
+        self._connection = sqlite3.connect(database)
+        self._connection.execute(
+            "CREATE TABLE IF NOT EXISTS tokens (id, refresh_token, updated_at)"
+        )
+        self._connection.execute(
+            "CREATE UNIQUE INDEX IF NOT EXISTS ux_tokens_id on tokens(id)"
+        )
+        self._connection.commit()
+        self.key = key
+
+    def _get(self):
+        cursor = self._connection.execute(
+            "SELECT refresh_token FROM tokens WHERE id=?", (self.key,)
+        )
+        result = cursor.fetchone()
+        if result is None:
+            raise KeyError
+        return result[0]
+
+    def _set(self, refresh_token):
+        """Set the refresh token in the database.
+
+        This function will overwrite an existing value if the corresponding ``key``
+        already exists.
+
+        """
+        self._connection.execute(
+            "REPLACE INTO tokens VALUES (?, ?, datetime('now'))",
+            (self.key, refresh_token),
+        )
+        self._connection.commit()
+
+    def is_registered(self):
+        """Return whether or not ``key`` already has a ``refresh_token``."""
+        cursor = self._connection.execute(
+            "SELECT refresh_token FROM tokens WHERE id=?", (self.key,)
+        )
+        return cursor.fetchone() is not None
+
+    def post_refresh_callback(self, authorizer):
+        """Update the refresh token in the database."""
+        self._set(authorizer.refresh_token)
+
+        # While the following line is not strictly necessary, it ensures that the
+        # refresh token is not used elsewhere. And also forces the pre_refresh_callback
+        # to always load the latest refresh_token from the database.
+        authorizer.refresh_token = None
+
+    def pre_refresh_callback(self, authorizer):
+        """Load the refresh token from the database."""
+        assert authorizer.refresh_token is None
+        authorizer.refresh_token = self._get()
+
+    def register(self, refresh_token):
+        """Register the initial refresh token in the database.
+
+        :returns: ``True`` if ``refresh_token`` is saved to the database, otherwise,
+            ``False`` if there is already a ``refresh_token`` for the associated
+            ``key``.
+
+        """
+        cursor = self._connection.execute(
+            "INSERT OR IGNORE INTO tokens VALUES (?, ?, datetime('now'))",
+            (self.key, refresh_token),
+        )
+        self._connection.commit()
+        return cursor.rowcount == 1